Refactor docker-compose to specify PostgreSQL version and update WeatherForecast model
